Committed dynamic version of the globe
CURRENT BUG:

The visualization slows down further and further as we switch between
datasets. This leads me to believe that there is something Three is
doing in the background in regards to how the canvas is being cleared
up and it’s causing it to keep the objects in the backend. Will fix
ASAP.